Change the Fundamental Identifier of Viewers
Using the Displayname as a Identifier clashes with Twitchtoolkit and creates a second user for any user which "login/username" doesn't match their "displayname" allowing them to basically duplicate their coingain and amount of pawns they can control. which can be abused and is a straight up inconvenience for people not abusing it.